+++ tests/tests/es.math.to-string-tag.js
@@ -1,0 +1,3 @@
+QUnit.test('Math[@@toStringTag]', assert => {
+  assert.strictEqual(Math[Symbol.toStringTag], 'Math', 'Math[@@toStringTag] is `Math`');
+});