+++ tests/tests/es.reflect.to-string-tag.js
@@ -1,0 +1,3 @@
+QUnit.test('Reflect[@@toStringTag]', assert => {
+  assert.strictEqual(Reflect[Symbol.toStringTag], 'Reflect', 'Reflect[@@toStringTag] is `Reflect`');
+});